requirements.txt
biopython==1.80
gemmi==0.5.8
numpy==1.24.1
